Set group trigger conditions
This change adds the functionality of setting group trigger conditions.
In case the group trigger exists it will check if an update is needed
for the existing conditions, and if so will replace them with the
new ones. The comparison is based on the condition name in its context
attribute.